requirements.txt
aiohttp == 3.8.6
akanekopy == 1.1.4
async-timeout == 4.0.3
beautifulsoup4 == 4.12.2
colorama == 0.4.6
comtypes == 1.2.0	
discord.py == 2.3.2
discord-together == 1.2.6
giphy_client == 1.0.0
lyricsgenius == 3.0.1
Pillow == 10.0.1
requests == 2.31.0
spotipy == 2.23.0
yarl == 1.9.2
youtube-search-python == 1.6.6
pylast == 5.2.0
aioschedule == 0.5.2
pymongo == 4.5.0 
google-api-python-client == 2.102.0
portalocker == 2.8.2
yt_dlp == 2023.10.7